Don't rely on the `pip` library in setup.py, it may break
The `pip` library never intended for their APIs to be used publicly, as detailed
in https://github.com/pypa/pip/issues/2286.

With pip 10 now available the `pip.req` module is no longer available and that
breaks installation of tornado-botocore on installations with pip 10 or newer.

Since the functionality of the library mostly was to read in dependencies
specified in `requirements.txt`, it's quite easy to fix this by just reading
the content of `requirements.txt` line by line and putting that into an array.